requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asarPy==1.0.1
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
Cython==0.29.37
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-bootstrap5==24.2
django-easy-select2==1.5.8
djangorestframework==3.15.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
fire==0.6.0
fqdn==1.5.1
fsspec==2024.3.1
fugashi==1.3.2
gpg==1.16.0
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
jaconv==0.3.4
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.2
louis==3.20.0
Mako==1.3.5
manga-ocr==0.1.11
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
meson==1.3.2
mistune==3.0.2
mokuro==0.1.8
more-itertools==8.10.0
mpmath==1.3.0
mutagen==1.47.0
natsort==8.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pulsemixer==1.5.1
pure-eval==0.2.2
pycairo==1.20.1
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodomex==3.20.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyperclip==1.8.2
python-apt===2.4.0-elementary9-ubuntu7.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.4.16
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
safetensors==0.4.3
scipy==1.13.0
SecretStorage==3.3.1
Send2Trash==1.8.3
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
triton==2.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unidic-lite==1.0.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
yattag==1.15.2
youtube-dl==2021.12.17
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
zipp==1.0.0
